+++ src/Physics/PHHapticEngineImpulse.cpp
@@ -47,7 +47,7 @@ void PHHapticEngineImpulse::SyncHaptic2Physic(){
 	for(int i = 0; i < hapticLoop->NHapticPointers(); i++){
 		PHHapticPointer* hpointer = hapticLoop->GetHapticPointer(i);
 		int hpointerID = hpointer->GetPointerID();
-		int nNeighbors = hpointer->neighborSolidIDs.size();
+		int nNeighbors = (int)hpointer->neighborSolidIDs.size();
 		// 近傍物体であるペアだけ同期
 		for(int j = 0; j < nNeighbors; j++){
 			int solidID = hpointer->neighborSolidIDs[j];
@@ -84,7 +84,7 @@ void PHHapticEngineImpulse::SyncPhysic2Haptic(){
 	for(int i = 0; i < NHapticPointers(); i++){
 		PHHapticPointer* ppointer = GetHapticPointer(i);
 		const int ppointerID = ppointer->GetPointerID();
-		const int nNeighbors = ppointer->neighborSolidIDs.size();
+		const int nNeighbors = (int)ppointer->neighborSolidIDs.size();
 		for(int j = 0; j < nNeighbors; j++){
 			const int solidID = ppointer->neighborSolidIDs[j];
 			PHSolidPairForHaptic* hpair = hapticLoop->GetSolidPairForHaptic(solidID, ppointerID);
